알고리즘 별 calc_bt 일치, tt/wt 정수로 변환

+++ src/pyodide-simulator/models.py
@@ -35,8 +35,8 @@ class ProcessMetric:
     pid: str
     bt: int
     at: int
-    tt: float
-    wt: float
+    tt: int
+    wt: int
     ntt: float
